backend/requirements.txt
bcrypt==4.3.0
# bcrypt==4.1.2
email-validator===2.2.0
fastapi==0.115.13
passlib==1.7.4
pydantic==2.11.7
pydantic-extra-types==2.10.5
prisma==0.15.0
pyjwt==2.10.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart===0.0.20
pymongo==4.13.2
yagmail==0.15.293
tzlocal==5.3.1
apscheduler==3.11.0
uvicorn[standard]==0.35.0
